requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
av==15.0.0
banks==2.2.0
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.3
cssselect==1.3.0
cssutils==2.11.1
ctranslate2==4.6.0
cycler==0.12.1
Cython==3.1.2
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
faster-whisper==1.2.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
GPUtil==1.4.0
greenlet==3.2.4
griffe==1.11.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.13
lazy_loader==0.4
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.0
llama-index-cli==0.5.0
llama-index-core==0.13.0
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.0
llama-index-instrumentation==0.4.0
llama-index-llms-openai==0.5.2
llama-index-readers-file==0.5.0
llama-index-readers-llama-parse==0.5.0
llama-index-workflows==1.3.0
llama-parse==0.6.54
lmdb==1.7.3
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
onnxruntime==1.22.1
openai==1.99.3
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.1
packaging==25.0
paddleocr==3.1.0
paddlepaddle==3.1.0
paddlex==3.1.3
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
protobuf==6.31.1
py-cpuinfo==9.0.0
PyAudio==0.2.14
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pygame==2.6.1
pyparsing==3.2.3
pypdf==5.9.0
pypdfium2==4.30.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
shapely==2.1.1
simsimd==6.5.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.42
stringzilla==3.12.6
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.10.0
tokenizers==0.21.4
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
